pyproject.toml
[project]
name = "embedding-teste"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psutil>=7.2.2",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.3",
]
